Refactor hostSettings name
